changed: put art videodatabase methods in separate file
