+++ xbmc/video/database/CMakeLists.txt
@@ -1,4 +1,5 @@
 set(SOURCES VideoDatabase.cpp
+            VideoDatabaseArt.cpp
             VideoDatabaseBookmarks.cpp
             VideoDatabaseExportImport.cpp
             VideoDatabaseMovies.cpp
